Move TestPodIPsIndexFunc to package controller/grouping
Move TestPodIPsIndexFunc from package controller/traceflow to controller/grouping,
because PodIPsIndexFunc is in package controller/grouping.
